Separating out the YTDL classes.
