+++ Jellyfin.Plugin.YoutubeMetadata/Providers/YoutubeDL/YTDLEpisodeImageProvider.cs
@@ -1,0 +1,109 @@
+﻿using System.Collections.Generic;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using MediaBrowser.Controller.Configuration;
+using MediaBrowser.Controller.Providers;
+using MediaBrowser.Model.IO;
+using MediaBrowser.Model.Providers;
+using Microsoft.Extensions.Logging;
+using MediaBrowser.Controller.Entities;
+using MediaBrowser.Controller.Entities.Movies;
+using MediaBrowser.Model.Entities;
+
+namespace Jellyfin.Plugin.YoutubeMetadata.Providers.YoutubeDL
+{
+    public class YTDLEpisodeImageProvider : IRemoteImageProvider, IHasOrder
+    {
+        private readonly IServerConfigurationManager _config;
+        private readonly IFileSystem _fileSystem;
+        private readonly IHttpClientFactory _httpClientFactory;
+        private readonly ILogger<YoutubeMetadataImageProvider> _logger;
+
+        public YTDLEpisodeImageProvider(IServerConfigurationManager config, IFileSystem fileSystem, IHttpClientFactory httpClientFactory, ILogger<YoutubeMetadataImageProvider> logger)
+        {
+            _config = config;
+            _fileSystem = fileSystem;
+            _httpClientFactory = httpClientFactory;
+            _logger = logger;
+        }
+
+        /// <summary>
+        /// Providers name, this appears in the library metadata settings.
+        /// </summary>
+        public string Name => "YouTube-DL Image Metadata";
+
+        /// <inheritdoc />
+        // After embedded and fanart
+        public int Order => 1;
+
+        /// <summary>
+        /// Gets the supported images.
+        /// </summary>
+        /// <param name="item"></param>
+        /// <returns></returns>
+        public IEnumerable<ImageType> GetSupportedImages(BaseItem item)
+        {
+            return new List<ImageType>
+            {
+                ImageType.Primary,
+                ImageType.Disc
+            };
+        }
+
+        /// <summary>
+        /// Retrieves image for item.
+        /// </summary>
+        /// <param name="item"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        public async Task<IEnumerable<RemoteImageInfo>> GetImages(BaseItem item, CancellationToken cancellationToken)
+        {
+            var result = new List<RemoteImageInfo>();
+            var id = Utils.GetYTID(item.FileNameWithoutExtension);
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                _logger.LogInformation("Youtube ID not found in filename of title: " + item.Name);
+                return result;
+            }
+            var ytPath = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
+            var fileInfo = _fileSystem.GetFileSystemInfo(ytPath);
+            if (!(Utils.IsFresh(fileInfo)))
+            {
+                await Utils.YTDLMetadata(id, _config.ApplicationPaths, cancellationToken);
+            }
+            var path = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
+            var video = Utils.ReadYTDLInfo(path, cancellationToken);
+            if (video != null)
+            {
+                result.Add(new RemoteImageInfo
+                {
+                    ProviderName = Name,
+                    Url = video.thumbnails[video.thumbnails.Count - 1].URL,
+                    Type = ImageType.Primary
+                });
+            }
+            return result;
+        }
+
+        /// <summary>
+        /// Gets the image response.
+        /// </summary>
+        /// <param name="url"></param>
+        /// <param name="cancellationToken"></param>
+        /// <returns></returns>
+        public Task<HttpResponseMessage> GetImageResponse(string url, CancellationToken cancellationToken)
+        {
+            var httpClient = _httpClientFactory.CreateClient();
+            return httpClient.GetAsync(url, cancellationToken);
+        }
+
+        /// <summary>
+        /// Returns True if item is supported.
+        /// </summary>
+        /// <param name="item"></param>
+        /// <returns></returns>
+        public bool Supports(BaseItem item)
+            => item is Movie;
+    }
+}
