+++ Jellyfin.Plugin.YoutubeMetadata/Providers/YoutubeDL/YTDLSeriesProvider.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using MediaBrowser.Controller.Providers;
+using MediaBrowser.Model.Providers;
+using MediaBrowser.Controller.Entities.TV;
+
+namespace Jellyfin.Plugin.YoutubeMetadata.Providers.YoutubeDL
+{
+    public class YTDLSeriesProvider : IRemoteMetadataProvider<Series, SeriesInfo>
+    {
+        public string Name => "YouTube-DL Series Metadata";
+
+        public Task<HttpResponseMessage> GetImageResponse(string url, CancellationToken cancellationToken)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<MetadataResult<Series>> GetMetadata(SeriesInfo info, CancellationToken cancellationToken)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<IEnumerable<RemoteSearchResult>> GetSearchResults(SeriesInfo searchInfo, CancellationToken cancellationToken)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
