--- Jellyfin.Plugin.YoutubeMetadata/Providers/YoutubeDL/YoutubeDLMetadataProvider.cs
@@ -1,281 +1,0 @@
-﻿using System;
-using System.Text.Json;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-using MediaBrowser.Controller.Configuration;
-using MediaBrowser.Controller.Library;
-using MediaBrowser.Controller.Providers;
-using MediaBrowser.Model.IO;
-using MediaBrowser.Model.Providers;
-using Microsoft.Extensions.Logging;
-using MediaBrowser.Controller.Entities;
-using MediaBrowser.Controller.Entities.Movies;
-using MediaBrowser.Model.Entities;
-using MediaBrowser.Controller.Entities.TV;
-
-namespace Jellyfin.Plugin.YoutubeMetadata.Providers
-{
-    public class YoutubeDLEpisodeProvider : IRemoteMetadataProvider<Episode, EpisodeInfo>, IHasOrder
-    {
-        private readonly IServerConfigurationManager _config;
-        private readonly IFileSystem _fileSystem;
-        private readonly ILogger<YoutubeMetadataProvider> _logger;
-
-        public const string BaseUrl = "https://m.youtube.com/";
-
-        public YoutubeDLEpisodeProvider(IServerConfigurationManager config, IFileSystem fileSystem, ILogger<YoutubeMetadataProvider> logger)
-        {
-            _config = config;
-            _fileSystem = fileSystem;
-            _logger = logger;
-        }
-        public string Name => "YouTube-DL Metadata";
-
-        public int Order => 1;
-
-        public Task<HttpResponseMessage> GetImageResponse(string url, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-
-        public async Task<MetadataResult<Episode>> GetMetadata(EpisodeInfo info, CancellationToken cancellationToken)
-        {
-            Console.WriteLine("Inside GetMetadata");
-            var result = new MetadataResult<Episode>();
-            
-            var id = Utils.GetYTID(info.Path);
-            if (string.IsNullOrWhiteSpace(id))
-            {
-                _logger.LogInformation("Youtube ID not found in filename of title: " + info.Name);
-                return result;
-            }
-            var ytPath = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
-            var fileInfo = _fileSystem.GetFileSystemInfo(ytPath);
-            if (!(Utils.IsFresh(fileInfo)))
-            {
-                await Utils.YTDLMetadata(id, _config.ApplicationPaths, cancellationToken);
-            }
-            var path = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
-            var video = Utils.ReadYTDLInfo(path, cancellationToken);
-            if (video != null)
-            {
-                result = Utils.YTDLJsonToEpisode(video);
-            }
-
-            return result;
-        }
-
-        public Task<IEnumerable<RemoteSearchResult>> GetSearchResults(EpisodeInfo searchInfo, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-
-
-    public class YoutubeDLSeriesProvider : IRemoteMetadataProvider<Series, SeriesInfo>
-    {
-        public string Name => "YouTube-DL Metadata";
-
-        public Task<HttpResponseMessage> GetImageResponse(string url, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<MetadataResult<Series>> GetMetadata(SeriesInfo info, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<IEnumerable<RemoteSearchResult>> GetSearchResults(SeriesInfo searchInfo, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class YoutubeDLMetadataProvider : IRemoteMetadataProvider<Movie, MovieInfo>, IHasOrder
-    {
-        private readonly IServerConfigurationManager _config;
-        private readonly IFileSystem _fileSystem;
-        private readonly ILogger<YoutubeMetadataProvider> _logger;
-        private readonly ILibraryManager _libmanager;
-
-        public const string BaseUrl = "https://m.youtube.com/";
-        
-        public YoutubeDLMetadataProvider(IServerConfigurationManager config, IFileSystem fileSystem, ILogger<YoutubeMetadataProvider> logger, ILibraryManager libmanager)
-        {
-            _config = config;
-            _fileSystem = fileSystem;
-            _logger = logger;
-            _libmanager = libmanager;
-        }
-
-        /// <summary>
-        /// Providers name, this appears in the library metadata settings.
-        /// </summary>
-        public string Name => "YouTube-DL Metadata";
-
-        /// <inheritdoc />
-        public int Order => 1;
-
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="searchInfo"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        public Task<IEnumerable<RemoteSearchResult>> GetSearchResults(MovieInfo searchInfo, CancellationToken cancellationToken)
-            => Task.FromResult(Enumerable.Empty<RemoteSearchResult>());
-
-        private string GetPathByTitle(string title)
-        {
-            var query = new InternalItemsQuery { Name = title };
-            var results = _libmanager.GetItemsResult(query);
-            return results.Items[0].Path;
-        }
-
-        /// <inheritdoc />
-        public async Task<MetadataResult<Movie>> GetMetadata(MovieInfo info, CancellationToken cancellationToken)
-        {
-            var result = new MetadataResult<Movie>();
-            var id = Utils.GetYTID(GetPathByTitle(info.Name));
-            if (string.IsNullOrWhiteSpace(id))
-            {
-                _logger.LogInformation("Youtube ID not found in filename of title: " + info.Name);
-                return result;
-            }
-            var ytPath = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
-            var fileInfo = _fileSystem.GetFileSystemInfo(ytPath);
-            if (!(Utils.IsFresh(fileInfo)))
-            {
-                await Utils.YTDLMetadata(id, _config.ApplicationPaths, cancellationToken);
-            }
-            var path = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
-            var video = Utils.ReadYTDLInfo(path, cancellationToken);
-            if (video != null)
-            {
-                result = Utils.YTDLJsonToMovie(video);
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Processes the found metadata into the Movie entity.
-        /// </summary>
-        /// <param name="item"></param>
-        /// <param name="result"></param>
-        /// <param name="preferredLanguage"></param>
-        public static void ProcessResult(Video item, Google.Apis.YouTube.v3.Data.Video result)
-        {
-            item.Name = result.Snippet.Title;
-            item.Overview = result.Snippet.Description;
-            var date = DateTime.Parse(result.Snippet.PublishedAtRaw);
-            item.ProductionYear = date.Year;
-            item.PremiereDate = date;
-        }
-
-        public Task<HttpResponseMessage> GetImageResponse(string url, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public class YTDLImageProvider : IRemoteImageProvider, IHasOrder
-    {
-        private readonly IServerConfigurationManager _config;
-        private readonly IFileSystem _fileSystem;
-        private readonly IHttpClientFactory _httpClientFactory;
-        private readonly ILogger<YoutubeMetadataImageProvider> _logger;
-
-        public YTDLImageProvider(IServerConfigurationManager config, IFileSystem fileSystem, IHttpClientFactory httpClientFactory, ILogger<YoutubeMetadataImageProvider> logger)
-        {
-            _config = config;
-            _fileSystem = fileSystem;
-            _httpClientFactory = httpClientFactory;
-            _logger = logger;
-        }
-
-        /// <summary>
-        /// Providers name, this appears in the library metadata settings.
-        /// </summary>
-        public string Name => "YouTube-DL Image Metadata";
-
-        /// <inheritdoc />
-        // After embedded and fanart
-        public int Order => 1;
-
-        /// <summary>
-        /// Gets the supported images.
-        /// </summary>
-        /// <param name="item"></param>
-        /// <returns></returns>
-        public IEnumerable<ImageType> GetSupportedImages(BaseItem item)
-        {
-            return new List<ImageType>
-            {
-                ImageType.Primary,
-                ImageType.Disc
-            };
-        }
-
-        /// <summary>
-        /// Retrieves image for item.
-        /// </summary>
-        /// <param name="item"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        public async Task<IEnumerable<RemoteImageInfo>> GetImages(BaseItem item, CancellationToken cancellationToken)
-        {
-            var result = new List<RemoteImageInfo>();
-            var id = Utils.GetYTID(item.FileNameWithoutExtension);
-            if (string.IsNullOrWhiteSpace(id))
-            {
-                _logger.LogInformation("Youtube ID not found in filename of title: " + item.Name);
-                return result;
-            }
-            var ytPath = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
-            var fileInfo = _fileSystem.GetFileSystemInfo(ytPath);
-            if (!(Utils.IsFresh(fileInfo)))
-            {
-                await Utils.YTDLMetadata(id, _config.ApplicationPaths, cancellationToken);
-            }
-            var path = Utils.GetVideoInfoPath(_config.ApplicationPaths, id);
-            var video = Utils.ReadYTDLInfo(path, cancellationToken);
-            if (video != null)
-            {
-                result.Add(new RemoteImageInfo
-                {
-                    ProviderName = Name,
-                    Url = video.thumbnails[video.thumbnails.Count - 1].URL,
-                    Type = ImageType.Primary
-                });
-            }
-            return result;
-        }
-
-        /// <summary>
-        /// Gets the image response.
-        /// </summary>
-        /// <param name="url"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        public Task<HttpResponseMessage> GetImageResponse(string url, CancellationToken cancellationToken)
-        {
-            var httpClient = _httpClientFactory.CreateClient();
-            return httpClient.GetAsync(url, cancellationToken);
-        }
-
-        /// <summary>
-        /// Returns True if item is supported.
-        /// </summary>
-        /// <param name="item"></param>
-        /// <returns></returns>
-        public bool Supports(BaseItem item)
-            => item is Movie;
-    }
-}
